+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### master
 
-- Fix a problem with quoted series names in legend
+#### Legend
+
+- Fix a problem with quoted series names
 
 ### 1.3.3 (11-10-2021)