+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### master
 
+#### Legend
+
+- Fix a problem with quoted series names
+
 ### 1.3.3 (11-10-2021)
 
 #### Tooltip